+++ src/game/player.rs
@@ -2,7 +2,7 @@
 use ::util::vector::Vector2D;
 
 pub struct Player {
-    position: Vector2D<i16>,
+    position: Vector2D<i32>,
 }
 
 impl Player {
@@ -12,11 +12,11 @@ impl Player {
         }
     }
 
-    pub fn position(&self) -> &Vector2D<i16> {
+    pub fn position(&self) -> &Vector2D<i32> {
         &self.position
     }
 
-    pub fn move_position(&mut self, x_move: i16, y_move: i16) {
+    pub fn move_position(&mut self, x_move: i32, y_move: i32) {
         self.position.x += x_move;
         self.position.y += y_move;
     }
