+++ src/graphics/renderer.rs
@@ -8,19 +8,19 @@ use super::sprite::Sprite;
 
 
 struct RenderSection {
-    start_point: Vector2D<u8>,
-    size: Vector2D<u8>,
+    start_point: Vector2D<i32>,
+    size: Vector2D<i32>,
 
 }
 
 pub struct Renderer {
-    size: Vector2D<u8>,
+    size: Vector2D<i32>,
     clear_colour: Colour,
     render_sections: HashMap<String, RenderSection>
 }
 
 impl RenderSection {
-    pub fn new(start_point: Vector2D<u8>, size: Vector2D<u8>) -> RenderSection {
+    pub fn new(start_point: Vector2D<i32>, size: Vector2D<i32>) -> RenderSection {
         RenderSection {
             start_point, 
             size,
@@ -29,7 +29,7 @@ impl RenderSection {
 }
 
 impl Renderer {
-    pub fn new(x_size: u8, y_size: u8) -> Renderer {
+    pub fn new(x_size: i32, y_size: i32) -> Renderer {
         let mut renderer = Renderer {
             size: Vector2D::new(x_size, y_size),
             clear_colour: Colour::new(25, 20, 70),
@@ -44,7 +44,7 @@ impl Renderer {
         renderer
     }
     
-    pub fn add_render_section(&mut self, name: &str, start_point: Vector2D<u8>, size: Vector2D<u8>) {
+    pub fn add_render_section(&mut self, name: &str, start_point: Vector2D<i32>, size: Vector2D<i32>) {
         self.render_sections.insert(
             name.to_string(), 
             RenderSection::new(start_point, size));
@@ -81,7 +81,7 @@ impl Renderer {
             Some(render_section) => {
                 for y in 0..render_section.size.y {
                     for x in 0..render_section.size.x {
-                        self.draw_string(section, " ", &Vector2D::new(x as i16, y as i16));
+                        self.draw_string(section, " ", &Vector2D::new(x, y));
                     }
                 }
             } 
@@ -92,7 +92,7 @@ impl Renderer {
         &self.clear_colour
     }
 
-    pub fn draw_solid_line_x(&self, colour: &Colour, begin_position: &Vector2D<u8>, length: u8) {
+    pub fn draw_solid_line_x(&self, colour: &Colour, begin_position: &Vector2D<i32>, length: i32) {
         Renderer::set_bg_colour(colour);
         Renderer::set_cursor_location(begin_position.x + 1, begin_position.y + 1);
         for _x in begin_position.x..length {
@@ -101,7 +101,7 @@ impl Renderer {
         Renderer::set_bg_colour(&self.clear_colour);
     }
 
-    pub fn draw_solid_line_y(&self, colour: &Colour, begin_position: &Vector2D<u8>, height: u8) {
+    pub fn draw_solid_line_y(&self, colour: &Colour, begin_position: &Vector2D<i32>, height: i32) {
         Renderer::set_bg_colour(colour);
         for y in begin_position.y..height {
             Renderer::set_cursor_location(begin_position.x + 1, begin_position.y + y + 1);
@@ -129,7 +129,7 @@ impl Renderer {
     /*
      * Misc ANSI commands
      */
-    pub fn set_cursor_location(x: u8, y: u8) {
+    pub fn set_cursor_location(x: i32, y: i32) {
         print!("\x1b[{};{}H", y + 1, x + 1);
     }
 
@@ -137,7 +137,7 @@ impl Renderer {
      * Public drawing interface
      * self is used to ensure these functions are only called on the object itself and not globally
      */
-    pub fn set_cursor_render_section(&self, section: &str, position: &Vector2D<u8>) {
+    pub fn set_cursor_render_section(&self, section: &str, position: &Vector2D<i32>) {
         match self.render_sections.get(section) {
             None => panic!(format!("Tried to render to section which doesn't exist: {}", section)),
             Some(section) => {
@@ -146,17 +146,17 @@ impl Renderer {
         }
     }
 
-    pub fn draw_string(&self, section: &str, string: &str, start_position: &Vector2D<i16>) {
+    pub fn draw_string(&self, section: &str, string: &str, start_position: &Vector2D<i32>) {
         let sect = self.render_sections.get(section).unwrap();
 
-        if start_position.y < 0 || start_position.y >= sect.size.y as i16 {
+        if start_position.y < 0 || start_position.y >= sect.size.y {
             return;
         }
 
-        self.set_cursor_render_section(section, &Vector2D::new(start_position.x as u8, start_position.y as u8));
+        self.set_cursor_render_section(section, &Vector2D::new(start_position.x, start_position.y));
         print!("{}", string);
     }
-
+/*
     pub fn draw_sprite(&self, section: &str, sprite: &Sprite) {
         let position = &sprite.position;
         self.set_cursor_render_section(section, &Vector2D::new(position.x as u8, position.y as u8));
@@ -168,12 +168,12 @@ impl Renderer {
             line_num += 1;
         }
     }
-
+*/
 
     /*
         Misc functions
     */
-    pub fn get_size(&self) -> &Vector2D<u8> {
+    pub fn get_size(&self) -> &Vector2D<i32> {
         &self.size
     }
 }
